+++ js/data.js
@@ -1,3 +1,13 @@
+import { generateRandomNumber, getRandomElem, getRandomId } from "./util";
+
+const PHOTOS_NUM = 25;
+const LIKES = {min: 15, max: 200};
+const ARRAY_ID = Array(PHOTOS_NUM).fill(false);
+const ARRAY_URL = Array(PHOTOS_NUM).fill(false);
+const NUM_OF_COMMENTS = 3;
+const MAX_ID = 25;
+
+
 const DESCRIPTIONS = [
     'Lovely day',
     'Nice weather',
@@ -22,4 +32,21 @@ const DESCRIPTIONS = [
     'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
   ];
 
-export {DESCRIPTIONS, NAMES, MESSAGES};
+const addComment = () => ({
+    id: generateRandomNumber(1, 10000),
+    avatar: `img/avatar-${generateRandomNumber(1, 6)}.svg`,
+    message: getRandomElem(MESSAGES),
+    name: getRandomElem(NAMES)
+});
+
+const createDescription = (id) => ({
+  id: id,
+  url: `photos/${id}.jpg`,
+  description: getRandomElem(DESCRIPTIONS),
+  likes: generateRandomNumber(15, 200),
+  comments: Array.from({length: generateRandomNumber(1, 6)}, addComment)
+});
+
+const createDescriptions = () => Array.from({length: PHOTOS_NUM}, createDescription);
+
+export {createDescriptions};
