+++ js/util.js
@@ -1,4 +1,3 @@
-
 const generateRandomNumber = (minimum, maximum) => {
     if (minimum >= maximum) {
       throw new Error('Max string is less or equals the minimum string');
@@ -12,23 +11,17 @@ const generateRandomNumber = (minimum, maximum) => {
   };
   
   // eslint-disable-next-line no-unused-vars
-  const isUnderMaximum = (string, maxLength) => maxLength >= (string.length);
+const isUnderMaximum = (string, maxLength) => maxLength >= (string.length);
   
-  const getRandomElem = (elements) => elements[generateRandomNumber(0, elements.length - 1)];
-  
-  const addComment = () => ({
-    id: generateRandomNumber(1, 10000),
-    avatar: `img/avatar-${generateRandomNumber(1, 6)}.svg`,
-    message: getRandomElem(MESSAGES),
-    name: getRandomElem(NAMES)
-  });
-  
-  const createDescription = (id) => ({
-    id: id,
-    url: `photos/${id}.jpg`,
-    description: getRandomElem(DESCRIPTIONS),
-    likes: generateRandomNumber(15, 200),
-    comments: Array.from({length: generateRandomNumber(1, 6)}, addComment)
-  });
+const getRandomElem = (elements) => elements[generateRandomNumber(0, elements.length - 1)];
+   
+const getRandomId = (arrayOfNum) => {
+  for (let i = 0; i < arrayOfNum.length; i++) {
+    if (!arrayOfNum[i]) {
+      arrayOfNum[i] = true;
+      return i + 1;
+    }
+  }
+}  
 
-export {generateRandomNumber, isUnderMaximum, getRandomElem, addComment, createDescription};
+export {generateRandomNumber, isUnderMaximum, getRandomElem, addComment, createDescription, getRandomId};
